+++ CHANGELOG.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented in this file.
 
  - Danish localization
 
+### Fixed
+
+ - Aura not displaying when token is outside the visible canvas
+
 ## [0.21.0] - 2020-06-13
 
 ### Added